# webapi/openapi.generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "dixcord-bot webapi",
    "version": "3.0.0",
    "description": "Internal REST API backing the dixcord-bot Discord bot, webview, and webadmin services. Every route requires a service-account JWT (admin, bot, or webview) except login and health."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Service-account JWT obtained from POST /api/auth/login."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "ok",
          "error"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing, invalid, or expired bearer token.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AuthErrorResponse"
            }
          }
        }
      },
      "ForbiddenRole": {
        "description": "Authenticated, but the account role is not permitted for this route.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AuthErrorResponse"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Unexpected server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "paths": {
    "/api/users/me": {
      "get": {
        "operationId": "getCurrentUser",
        "tags": [
          "Users"
        ],
        "summary": "Current user profile",
        "responses": {
          "200": {
            "description": "The authenticated user's profile.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "email": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "role": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenRole"
          }
        }
      },
      "put": {
        "operationId": "updateCurrentUser",
        "tags": [
          "Users"
        ],
        "summary": "Update current user profile",
        "description": "Service accounts (admin/bot/webview) cannot update their own profile via this route.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated profile.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "email": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No valid fields to update.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Service account profiles are managed via environment configuration.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "operationId": "deleteCurrentUser",
        "tags": [
          "Users"
        ],
        "summary": "Delete current user account",
        "description": "Service accounts (admin/bot/webview) cannot be deleted via this route.",
        "responses": {
          "200": {
            "description": "Account deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Service accounts cannot be deleted via the API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/api/trigger-responses": {
      "get": {
        "operationId": "listTriggerResponses",
        "tags": [
          "Trigger Responses"
        ],
        "summary": "List all trigger-response pairs",
        "description": "Flat list of every trigger/response junction row, joined with its trigger and response text.",
        "responses": {
          "200": {
            "description": "All trigger-response pairs.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "triggerResponses": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "Junction (trigger_response) row id"
                          },
                          "trigger_id": {
                            "type": "integer"
                          },
                          "response_id": {
                            "type": "integer"
                          },
                          "trigger_string": {
                            "type": "string"
                          },
                          "response_string": {
                            "type": "string"
                          },
                          "response_order": {
                            "type": "integer",
                            "nullable": true
                          },
                          "weight": {
                            "type": "integer",
                            "nullable": true,
                            "minimum": 0,
                            "maximum": 100
                          },
                          "response_function": {
                            "type": "string",
                            "nullable": true,
                            "description": "Resolved trigger_response_functions.function_name for the row's response_function FK",
                            "or null.": null
                          },
                          "response_function_parameters": {
                            "type": "object",
                            "nullable": true,
                            "description": "JSON payload passed to response_function at dispatch time",
                            "or null.": null
                          },
                          "selection_mode": {
                            "type": "string",
                            "enum": [
                              "random",
                              "ordered",
                              "weighted"
                            ]
                          },
                          "created_at": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "operationId": "createTriggerResponseLink",
        "tags": [
          "Trigger Responses"
        ],
        "summary": "Create a single trigger-response pair",
        "description": "Requires the admin role. Convenience one-shot alternative to POST /api/trigger-responses/triggers: gets-or-creates the trigger (by trigger_string) and the response (by response_string), then links them with a new trigger_response row. Setting selection_mode here updates the trigger's mode.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "trigger_string",
                  "response_string"
                ],
                "properties": {
                  "trigger_string": {
                    "type": "string"
                  },
                  "response_string": {
                    "type": "string"
                  },
                  "response_order": {
                    "type": "integer",
                    "nullable": true
                  },
                  "selection_mode": {
                    "type": "string",
                    "enum": [
                      "random",
                      "ordered",
                      "weighted"
                    ],
                    "default": "random"
                  },
                  "weight": {
                    "type": "integer",
                    "nullable": true,
                    "minimum": 0,
                    "maximum": 100
                  },
                  "response_function": {
                    "type": "string",
                    "nullable": true,
                    "description": "Optional function key to dispatch to instead of replying with response_string; usable regardless of selection_mode."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created trigger-response link.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "id": {
                      "type": "integer",
                      "description": "Junction (trigger_response) row id"
                    },
                    "trigger_id": {
                      "type": "integer"
                    },
                    "response_id": {
                      "type": "integer"
                    },
                    "trigger_string": {
                      "type": "string"
                    },
                    "response_string": {
                      "type": "string"
                    },
                    "response_order": {
                      "type": "integer",
                      "nullable": true
                    },
                    "weight": {
                      "type": "integer",
                      "nullable": true,
                      "minimum": 0,
                      "maximum": 100
                    },
                    "response_function": {
                      "type": "string",
                      "nullable": true,
                      "description": "Resolved trigger_response_functions.function_name for the row's response_function FK",
                      "or null.": null
                    },
                    "response_function_parameters": {
                      "type": "object",
                      "nullable": true,
                      "description": "JSON payload passed to response_function at dispatch time",
                      "or null.": null
                    },
                    "selection_mode": {
                      "type": "string",
                      "enum": [
                        "random",
                        "ordered",
                        "weighted"
                      ]
                    },
                    "created_at": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenRole"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/api/pin-quips/random": {
      "get": {
        "operationId": "getRandomPinQuip",
        "tags": [
          "Pin Quips"
        ],
        "summary": "Get one random pin quip",
        "responses": {
          "200": {
            "description": "A random pin quip.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "quip": {
                      "type": "string"
                    },
                    "id": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "No pin quips exist yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/api/message-processing/pin-check": {
      "post": {
        "operationId": "checkPinStatus",
        "tags": [
          "Message Processing"
        ],
        "summary": "Check whether a message is already logged as pinned",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "messageId"
                ],
                "properties": {
                  "messageId": {
                    "type": "string",
                    "description": "Discord message snowflake."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pin status.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "alreadyPinned": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  }
}